requirements.txt
fastapi
uvicorn
schedule
pyyaml
requests
